backend/requirements/base.txt
Django==4.2.15
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
psycopg[binary]==3.2.1
celery==5.4.0
redis==5.1.0
channels==4.1.0
channels-redis==4.2.0
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
uvicorn[standard]==0.30.6
gunicorn==23.0.0
django-storages[s3]==1.14.4
django-waffle==4.1.0
sentry-sdk==2.12.0
python-decouple==3.8
